[refactor] user 4계층으로 변경
