+++ user/src/main/java/com/sparta/rooibos/user/application/auditing/AuditorAwareImpl.java
@@ -1,4 +1,4 @@
-package com.sparta.rooibos.user.infrastructure.auditing;
+package com.sparta.rooibos.user.application.auditing;
 
 import org.springframework.data.domain.AuditorAware;
 import org.springframework.stereotype.Component;
