+++ user/src/main/java/com/sparta/rooibos/user/application/service/core/UserApproveServiceImpl.java
@@ -5,7 +5,7 @@ import com.sparta.rooibos.user.application.exception.custom.UserErrorCode;
 import com.sparta.rooibos.user.application.service.port.UserApproveService;
 import com.sparta.rooibos.user.domain.entity.User;
 import com.sparta.rooibos.user.domain.repository.UserRepository;
-import com.sparta.rooibos.user.infrastructure.auditing.UserAuditorContext;
+import com.sparta.rooibos.user.application.auditing.UserAuditorContext;
 import lombok.RequiredArgsConstructor;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.stereotype.Service;
