+++ user/src/main/java/com/sparta/rooibos/user/infrastructure/security/CacheConfig.java
@@ -1,6 +1,6 @@
 package com.sparta.rooibos.user.infrastructure.security;
 
-import com.sparta.rooibos.user.infrastructure.auditing.UserAuditorContext;
+import com.sparta.rooibos.user.application.auditing.UserAuditorContext;
 import org.springframework.cache.interceptor.KeyGenerator;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
