+++ user/src/main/java/com/sparta/rooibos/user/infrastructure/web/AuditorHeaderInterceptor.java
@@ -1,6 +1,6 @@
 package com.sparta.rooibos.user.infrastructure.web;
 
-import com.sparta.rooibos.user.infrastructure.auditing.UserAuditorContext;
+import com.sparta.rooibos.user.application.auditing.UserAuditorContext;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.stereotype.Component;
