+++ user/src/main/java/com/sparta/rooibos/user/presentation/controller/MasterController.java
@@ -6,7 +6,7 @@ import com.sparta.rooibos.user.application.dto.request.UserUpdateRequest;
 import com.sparta.rooibos.user.application.dto.response.UserListResponse;
 import com.sparta.rooibos.user.application.dto.response.UserResponse;
 import com.sparta.rooibos.user.application.service.port.MasterService;
-import com.sparta.rooibos.user.infrastructure.aop.MasterOnlyCheck;
+import com.sparta.rooibos.user.application.aop.MasterOnlyCheck;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
